requirements.txt
pandas-2.2.3
asyncio-3.4.3
json5-0.10.0
typing-3.7.4.3
parsel-1.9.1
loguru-0.7.3
httpx 
